feat: implement JSON-LD extraction and URL sanitization helpers for Luma events
